feat : добавил реализацию через память
