+++ src/main/java/ru/yandex/practicum/filmorate/storage/inmemoryimpl/InMemoryFilmStorage.java
@@ -1,21 +1,34 @@
 package ru.yandex.practicum.filmorate.storage.inmemoryimpl;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.FilmNotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.interfaces.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.interfaces.LikeStorage;
+import ru.yandex.practicum.filmorate.storage.interfaces.UserStorage;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Component("InMemoryFilmStorage")
 @Slf4j
-public class InMemoryFilmStorage implements FilmStorage {
+public class InMemoryFilmStorage implements FilmStorage, LikeStorage {
 
     final HashMap<Integer, Film> films = new HashMap<>();
     private int id = 0;
+    private final HashMap<Integer, ArrayList<Integer>> filmLikes = new HashMap<>();
+    private final UserStorage userStorage;
+
+    @Autowired
+    public InMemoryFilmStorage(@Qualifier("InMemoryUserStorage") UserStorage userStorage) {
+        this.userStorage = userStorage;
+    }
 
     @Override
     public Film addFilm(Film film) {
@@ -46,13 +59,17 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public List<Film> getPopularFilms(int count) {
-        //я не могу вернуть реализацию через память, из-за того, что теперь для этого мне нужно, чтобы
-        //этот класс вызывал класс с лайками и наоборот, а спринг тупо не позволяет мне таких вольностей
-        //большая часть ребят вообще поудаляла старые реализации, пощади меня, я не хочу в академ
-        //я могу обойти эту проблему , добавив список лайков в поле класса Film, но в этом нет никакого абсолютно смысла,
-        //ведь эта реализация больше не основная, а в ответах на запросы будет лишнее поле, которое абсолютно лишнее по сути
-        //ну или преренести этот метод в интерфейс, LikeStorage, но я тогда выброшусь из окна
-        return null;
+        List<ArrayList<Integer>> result = new ArrayList<>();
+        for (Map.Entry<Integer, ArrayList<Integer>> film : filmLikes.entrySet()) {
+            ArrayList<Integer> likes = new ArrayList<>();
+            likes.add(film.getKey());
+            result.add(film.getValue());
+        }
+        return result.stream()
+                .limit(count)
+                .map(e -> e.get(e.size() - 1))
+                .map(films::get)
+                .collect(Collectors.toList());
     }
 
     @Override
@@ -60,6 +77,25 @@ public class InMemoryFilmStorage implements FilmStorage {
         if (!films.containsKey(id)) {
             throw new FilmNotFoundException(String.format("Фильм с id=%d не найден", id));
         }
+    }
+
+    @Override
+    public String like(int filmId, int userId) {
+        filmExistenceCheck(filmId);
+        userStorage.userExistenceCheck(userId);
+        if (!filmLikes.containsKey(filmId)) {
+            filmLikes.put(filmId, new ArrayList<>());
+        }
+        filmLikes.get(filmId).add(userId);
+        return "Лайк добавлен";
+    }
+
+    @Override
+    public String removeLike(int filmId, int userId) {
+        filmExistenceCheck(filmId);
+        userStorage.userExistenceCheck(userId);
+        filmLikes.get(filmId).remove((Integer) userId);
+        return "Лайк успешно удален";
     }
 
     private int newId() {
